app_python/requirements.txt
django==3.2.6
djangorestframework==3.12.4
uwsgi==2.0.19.1